VideoPlayer: fix and cleanup ffmpeg sw deinterlacing
